Created an Oculus Button Press Instruction System Mouckup.
The system still needs to think better how to show and hide the controller and hand.
Check why haptics is not working.
Think about using events or leave as it is with reference to controller.
Set the scene objects that will make the controllers hide and the message dissapear.
Animate controller buttons.